Move static field from Feline to AllowancesTransformer
We want to avoid a situation where we get a static initialization
deadlock as the `Feline` code hasn't finished its static block, but
there's already a code calling of its methods and touching the previous
static field.

Moving this static field outside remove this possibility as the
different classes don't share the clinit synchronization.